Move gateway meta utils to payment module
